Update STM32f4 driver and system files.

+++ SDK/platform/stm32f4xx/sys/sysdelay.c
@@ -8,7 +8,7 @@
 #include "sysdelay.h"
 #include "sys/include/core_cm4.h"
 
-extern unsigned long SystemCoreClock;
+extern unsigned int SystemCoreClock;
 /******************************************************************************
 **                      INTERNAL VARIABLE DEFINITIONS
 *******************************************************************************/
